+++ opentelemetry-sdk/tests/trace/test_implementation.py
@@ -29,11 +29,13 @@ class TestTracerImplementation(unittest.TestCase):
     def test_tracer(self):
         tracer = trace.TracerProvider().get_tracer(__name__)
         with tracer.start_span("test") as span:
-            self.assertNotEqual(span.get_context(), INVALID_SPAN_CONTEXT)
+            self.assertNotEqual(span.get_span_context(), INVALID_SPAN_CONTEXT)
             self.assertNotEqual(span, INVALID_SPAN)
             self.assertIs(span.is_recording(), True)
             with tracer.start_span("test2") as span2:
-                self.assertNotEqual(span2.get_context(), INVALID_SPAN_CONTEXT)
+                self.assertNotEqual(
+                    span2.get_span_context(), INVALID_SPAN_CONTEXT
+                )
                 self.assertNotEqual(span2, INVALID_SPAN)
                 self.assertIs(span2.is_recording(), True)
 
@@ -43,5 +45,5 @@ class TestTracerImplementation(unittest.TestCase):
             span = trace._Span()
 
         span = trace._Span("name", INVALID_SPAN_CONTEXT)
-        self.assertEqual(span.get_context(), INVALID_SPAN_CONTEXT)
+        self.assertEqual(span.get_span_context(), INVALID_SPAN_CONTEXT)
         self.assertIs(span.is_recording(), True)