+++ opentelemetry-sdk/CHANGELOG.md
@@ -18,6 +18,8 @@
   ([#1162](https://github.com/open-telemetry/opentelemetry-python/pull/1162))
 - Adding Resource to MeterRecord
   ([#1209](https://github.com/open-telemetry/opentelemetry-python/pull/1209))
+- Parent is now always passed in via Context, intead of Span or SpanContext
+  ([#1146](https://github.com/open-telemetry/opentelemetry-python/pull/1146))
 
 ## Version 0.13b0